feat: product 모듈 분리(springBootTest configuration 문제있음)
