+++ app/src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,6 +3,9 @@ package kitchenpos.menu.application;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.application.dto.ProductInformationRequest;
+import kitchenpos.domain.Product;
+import kitchenpos.domain.repository.ProductRepository;
 import kitchenpos.menu.application.dto.MenuProductRequest;
 import kitchenpos.menu.application.dto.MenuRequest;
 import kitchenpos.menu.application.dto.MenuResponse;
@@ -13,9 +16,6 @@ import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menu.domain.repository.MenuProductRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
-import kitchenpos.product.application.dto.ProductInformationRequest;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
