+++ app/src/main/java/kitchenpos/menu/domain/MenuProductEventHandler.java
@@ -1,9 +1,9 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.element.Price;
+import kitchenpos.domain.MenuProductEvent;
+import kitchenpos.domain.Price;
 import kitchenpos.menu.domain.repository.MenuProductRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.product.domain.MenuProductEvent;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
