+++ product/src/main/java/kitchenpos/application/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.application.dto;
+package kitchenpos.application.dto;
 
 public class ProductRequest {
     private String name;
